refactor(type_checker): extract load_module into modules.rs
